django_service/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
python-dotenv
